requirements.txt
bleak==2.0.0
keyboard==0.13.5
pyserial==3.5
sep==1.4.1
